\name{birds_aminoAcids}
\alias{birds_aminoAcids}
\docType{data}
\title{
Bird Amino Acid Sequences
}
\description{
Data set containing amino acid sequences for 11,333 bird barcodes.
}
\usage{data(birds_aminoAcids)}
\format{
  The format is:
 chr [1:11333, 1:218] "gi|359280039|gb|JQ173884.1|" ...
 - attr(*, "dimnames")=List of 2
  ..$ : NULL
  ..$ : chr [1:218] "V1" "V2" "V3" "V4" ...
}
\source{
	Stoeckle, M. Y. and Kerr, K. C. R. (2012) Frequency Matrix Approach Demonstrates High Sequence Quality in Avian BARCODEs and Highlights Cryptic Pseudogenes. PLoS ONE. 7, e43992.
}
\examples{
\dontrun{data(birds_aminoAcids)}
}
\keyword{datasets}